requirements.txt
PyPDF2
Pillow
pymupdf
pytesseract
opencv-python
numpy
pdf2image
pandas
scikit-learn
requests
openai
ocrmypdf
img2pdf
pypdf
python-dotenv